IOUtils.jl

IOUtils.horizontal_line

IOUtils.indent

IOUtils.indent

IOUtils.print_boxed

IOUtils.print_boxed

IOUtils.redirect_output

IOUtils.@display

IOUtils.horizontal_line — Function

horizontal_line([io=stdout; char="━", color=:light_black])

Draw a colored horizontal line of chars across the whole screen.

Example

julia> horizontal_line()
━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

julia> horizontal_line(char="─")
────────────────────────────────────────────────────────────────────────────────

IOUtils.indent — Method

indent(fun, io, n[; indent_first=true])

Print all the output of fun to io indented by n spaces. If !indent_first, the first line is not indented.

Examples

julia> indent(stdout, 6) do io
           println(io, "Hello")
           println(io, "World")
       end
      Hello
      World

IOUtils.indent — Method

indent(fun, io, first_line::String)

Variant of indent that will first print first_line and then print all the output from fun indented by the length of first_line.

Examples

julia> indent(stdout, "Important information: ") do io
           println(io, "Hello")
           println(io, "World")
       end
Important information: Hello
                       World

julia> indent(stdout, "Important information: ") do io
           print_boxed(io) do io
               println(io, "Hello")
               println(io, "World")
           end
       end

Important information: ┌  Hello
                       └  World

IOUtils.print_boxed — Function

print_boxed(io, msglines[, prefix="", suffix=""; color=:light_black, chars="[┌│└"])

Print each line in msglines to io with an enclosing box drawn to the left, using the specified color. The first and last line will additionally print the optional parameter prefix and suffix, respectively. The box drawing characters can be customized by setting the string chars to any combination of four characters.

Examples

julia> print_boxed(stdout, ["Hello", "World"], ">>>", "<<<")
┌ >>>  Hello
└  World <<<

julia> print_boxed(stdout, ["Hello world"], chars="(╭│╰")
(  Hello world

julia> print_boxed(stdout, ["Hello", "world"], chars="(╭│╰")
╭  Hello
╰  world

IOUtils.print_boxed — Method

print_boxed(fun::Function, io, args...; kwargs...)

Block-version of print_boxed that via redirect_output captures the output of fun, and prints them in a block.

Examples

julia> print_boxed(stdout) do io
           println(io, "Hello")
           println(io, "World")
       end
┌  Hello
└  World

IOUtils.redirect_output — Method

redirect_output(fun)

Create an IOBuffer, pass it to fun, and return the captured output as a string.

Examples

julia> redirect_output() do io
           println(io, "Hello")
           println(io, "World")
       end
"Hello\nWorld\n"

IOUtils.@display — Macro

@display expr

Display expr using print_boxed along with the location from where the macro was called (useful for debugging); a mixture of @show, display, and @debug.

Examples

julia> @display sin.(1:10)
┌ sin.(1:10) =  10-element Array{Float64,1}:
│    0.8414709848078965
│    0.9092974268256817
│    0.1411200080598672
│   -0.7568024953079282
│   -0.9589242746631385
│   -0.27941549819892586
│    0.6569865987187891
│    0.9893582466233818
│    0.4121184852417566
│   -0.5440211108893698
└  @ REPL[16]:1